+++ Editor/Interfaces/IUtilityNode.cs
@@ -1,5 +1,8 @@
 namespace NewGraph {
     public interface IUtilityNode {
+
+        bool ShouldColorizeBackground();
+
         bool CreateInspectorUI();
         bool CreateNameUI();
         void Initialize(NodeController nodeController);